--- aver/aver/Models/HistoryViewModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace aver.Models
-{
-    public class HistoryViewModel
-    {
-        public List<HistoryItem> Items { get; set; }
-    }
-}
