--- aver/aver/Views/History/Index.cshtml
@@ -1,14 +1,0 @@
-﻿@model aver.Models.HistoryViewModel
-
-@{
-    ViewBag.Title = "Index";
-}
-
-<h2>Index</h2>
-
-
-@foreach (var item in Model.Items)
-{
-    <h2>@item.Title</h2>
-
-}
