+++ aver/aver/Views/Settings/Index.cshtml
@@ -5,13 +5,7 @@
 
 
 <div class="container">
-    <h2>Stillingar</h2>
-    <div>
-        <ul>
-            <li>Fá sendar tilkynningar @Html.CheckBox("GetNotifications")</li>
-            <li>Flottasta stillingin @Html.CheckBox("GetNotifications")</li>
-        </ul>
+   
 
-    </div>
 
 </div>
