+++ aver/aver/Views/Shared/_Layout.cshtml
@@ -61,7 +61,7 @@
                     <a class="nav-link waves-effect waves-light" href="Booking">Bóka</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link waves-effect waves-light" href="History">Saga</a>
+                    <a class="nav-link waves-effect waves-light" href="User">About</a>
                 </li>
                 
                 @*<li class="nav-item dropdown">
