+++ aver/aver/Views/User/Index.cshtml
@@ -1,0 +1,41 @@
+﻿@model aver.Models.UserViewModel
+
+@{
+    ViewBag.Title = "About you";
+}
+
+
+
+
+
+<!-- Card -->
+<div class="card">
+
+    <!-- Card image -->
+    <div class="view overlay">
+        <img class="card-img-top" src="https://mdbootstrap.com/img/Mockups/Lightbox/Thumbnail/img%20(67).jpg" alt="Card image cap">
+        <a href="#!">
+            <div class="mask rgba-white-slight"></div>
+        </a>
+    </div>
+
+    <!-- Card content -->
+    <div class="card-body">
+
+        <!-- Title -->
+        <h4 class="card-title">About you</h4>
+        <!-- Text -->
+        <p class="card-text">This is what we know about you</p>
+        <!-- Button -->
+        @foreach (var item in Model.Items)
+        {
+            <p class="card-text">@item.Title</p>
+
+        }
+
+    </div>
+
+</div>
+<!-- Card -->
+
+
